backend/requirements.txt
Django==3.2.7
djangorestframework==3.12.4
PyJWT==2.3.0
python-dotenv==0.19.1
